requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-generativeai==0.8.3
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
loguru==0.7.3
slowapi==0.1.9